+++ src/cone/app/browser/layout.py
@@ -9,7 +9,9 @@ from cone.app.browser.utils import make_query
 from cone.app.browser.utils import make_url
 from cone.app.browser.utils import node_icon
 from cone.app.browser.utils import node_path
+from cone.app.interfaces import IApplicationNode
 from cone.app.interfaces import ILayout
+from cone.app.interfaces import INavigationLeaf
 from cone.app.interfaces import IWorkflowState
 from cone.app.model import AppRoot
 from cone.app.ugm import principal_data
@@ -235,12 +237,11 @@ class MainMenu(LayoutConfigTile):
         # XXX: icons
         for key in root.keys():
             child = root[key]
-            props = child.properties
-            if self.ignore_node(child, props):
+            if self.ignore_node(child):
                 continue
             selected = curpath == key
-            item = self.create_item(child, props, empty_title, selected)
-            if props.mainmenu_display_children:
+            item = self.create_item(child, empty_title, selected)
+            if child.properties.mainmenu_display_children:
                 item['children'] = self.create_children(child, selected)
             else:
                 item['children'] = None
@@ -256,22 +257,23 @@ class MainMenu(LayoutConfigTile):
             curpath = ''
         for key in node.keys():
             child = node[key]
-            props = child.properties
-            if self.ignore_node(child, props):
+            if self.ignore_node(child):
                 continue
             selected = curpath == key
-            item = self.create_item(child, props, False, selected)
+            item = self.create_item(child, False, selected)
             children.append(item)
         return children
 
-    def ignore_node(self, node, props):
-        if props.skip_mainmenu:
+    def ignore_node(self, node):
+        if not IApplicationNode.providedBy(node):
+            return True
+        if node.properties.skip_mainmenu:
             return True
         if not self.request.has_permission('view', node):
             return True
         return False
 
-    def create_item(self, node, props, empty_title, selected):
+    def create_item(self, node, empty_title, selected):
         md = node.metadata
         item = dict()
         item['id'] = node.name
@@ -282,7 +284,7 @@ class MainMenu(LayoutConfigTile):
             item['title'] = md.title
             item['description'] = md.description
         item['url'] = make_url(self.request, node=node)
-        query = make_query(contenttile=props.default_content_tile)
+        query = make_query(contenttile=node.properties.default_content_tile)
         item['target'] = make_url(self.request, node=node, query=query)
         item['selected'] = selected
         item['icon'] = node_icon(node)
@@ -380,8 +382,8 @@ class NavTree(Tile):
         return item
 
     def fillchildren(self, model, path, tree):
-        """XXX: consider cone.app.interfaces.INavigationLeaf
-        """
+        if INavigationLeaf.providedBy(model):
+            return
         curpath = None
         if path:
             curpath = path[0]
@@ -401,9 +403,11 @@ class NavTree(Tile):
                 curpath = model.properties.default_child
         for key in model:
             node = model[key]
+            if not IApplicationNode.providedBy(node):
+                continue
             if not self.request.has_permission('view', node):
                 continue
-            if not node.properties.get('in_navtree'):
+            if not node.properties.in_navtree:
                 continue
             title = node.metadata.title
             if title: